fix: data type of amount

+++ database/schema.sql
@@ -9,7 +9,7 @@ CREATE TABLE IF NOT EXISTS drivers (
     password VARCHAR(100) NOT NULL,
     cnh VARCHAR(20) PRIMARY KEY NOT NULL,
     qrcode VARCHAR(100) NOT NULL,
-    amount BIGINT NOT NULL,
+    amount NUMERIC(10, 2) NOT NULL,
     street VARCHAR(100) NOT NULL,
     number VARCHAR(10) NOT NULL,
     complement VARCHAR(10),
